Set uniforms by name, and delete what nothing calls
Shader::set_uniform_block existed for braced-list call syntax and for
zeroing the padding member a block needs. The dotted name path,
set_uniform("fsp.gain", v), already worked on both backends, so the block
form was a type-erased UniformValue with a 64-byte buffer, two backend
implementations of block_member_names(), and about 110 lines to save one
set_uniform("vsp._pad", 0). The _pad member itself stays: without an int
in the block, SPIRV-Cross flattens it and the names disappear.

Gone with it, all of them unreferenced: LogOnce, ipc_server(),
prepend_includes(), the .glsl and .metal entries in shader_extensions and
the asset globs that matched them, and bench_heif_decode.cpp, which
answered the dav1d question that 323d37e9 settled and which no CI job
builds. describe_search_locations() is the search-path list again, now
that check-appimage.sh catches the packaging bug it was written for.

The IPC packet builders were production code that only the tests used,
since HDRView only ever receives; they live in tests/ipc_test_helpers.h.
The four test targets shared eight lines of setup written out four times,
now hdrview_setup_target().

326 cases, 361,732 assertions; hdrview_gui_tests 101/101. The Metal paths
here compile nowhere on this machine and are read, not built.
